Refactoring request validator to seperate classes
